requirements.txt
-i https://mirrors.aliyun.com/pypi/simple/
Django==3.1.4
django-bootstrap-form==3.4
django-bootstrap-pagination==1.7.1
requests==2.25.0
django-markdown-deux==1.0.5
markdown2==2.3.10
python-docx==0.8.10
lxml==4.6.2
# mysqlclient==1.4.6
# pymysql==0.10.1
gunicorn[gevent]==20.0.4
django-cors-headers==3.10.1
cryptography==42.0.2
